requirements.txt

fastapi
pydantic
python-dotenv